feat: improve gen meta fn
