+++ src/app/_podcast/page.tsx
@@ -1,6 +1,12 @@
 import NoData from "@/components/NoData";
 import { Container } from "@mui/material";
 import { PAGE_MIN_HEIGHT } from "@/constants";
+import { genMeta } from "@/utils/route";
+
+export const generateMetadata = genMeta(() => ({
+  titleSuffix: "Podcast",
+  relativeURL: "podcast",
+}));
 
 const PodcastPage = () => {
   return (
