--- src/app/challenges/[slug]/layout.tsx
@@ -1,39 +1,0 @@
-import { Stack } from "@mui/material";
-
-import { genMeta } from "@/utils/route";
-import BackLink from "@/components/Back";
-import PageHeaderWrapper from "@/components/PageHeaderWrapper";
-// import MoreContent from "./MoreContent";
-
-import ChallengeHeader from "./ChallengeHeader";
-import ChallengeNavigation from "./ChallengeNavigation";
-import markdownData from "../markdownData.json";
-
-export async function generateMetadata({ params: { slug } }) {
-  return genMeta({ relativeURL: `/challenges/${slug}` });
-}
-
-export default async function ChallengeDetailPage({ params, children }) {
-  const data = markdownData.find((x) => x.id === params.slug);
-  return (
-    <>
-      <PageHeaderWrapper
-        bgColor="rgba(255, 238, 218, 0.30)"
-        sx={{ pt: ["60px", "60px", "84px"], pb: ["30px", "40px", "60px"] }}
-      >
-        <Stack>
-          <BackLink></BackLink>
-          <ChallengeHeader
-            sx={{ mt: ["56px", "56px", "60px"] }}
-            challengeData={data}
-          />
-        </Stack>
-      </PageHeaderWrapper>
-
-      {children}
-
-      {/* <MoreContent /> */}
-      <ChallengeNavigation challengeId={params.slug}></ChallengeNavigation>
-    </>
-  );
-}
