+++ src/app/challenges/layout.tsx
@@ -1,11 +1,9 @@
 import { genMeta } from "@/utils/route";
 
-export async function generateMetadata() {
-  return genMeta({
-    titleSuffix: "Challenges",
-    relativeURL: "/challenges",
-  });
-}
+export const generateMetadata = genMeta(() => ({
+  titleSuffix: "Challenges",
+  relativeURL: "/challenges",
+}));
 
 export default function ChallengesLayout({ children }) {
   return <>{children}</>;
