+++ src/app/content/layout.tsx
@@ -1,8 +1,9 @@
 import { genMeta } from "@/utils/route";
 
-export async function generateMetadata() {
-  return genMeta({ titleSuffix: "Content", relativeURL: "/content" });
-}
+export const generateMetadata = genMeta(() => ({
+  titleSuffix: "Content",
+  relativeURL: "/content",
+}));
 
 const ContentLayout = async ({ children }) => {
   return <>{children}</>;
