+++ src/app/events/[slug]/register/page.tsx
@@ -1,11 +1,17 @@
 import TallyForm from "@/components/TallyForm";
 import Data from "../../eventsList.json";
+import { genMeta } from "@/utils/route";
 
 export function generateStaticParams() {
   return Data.map((event) => ({
     params: { slug: event.url },
   }));
 }
+
+export const generateMetadata = genMeta(({ params: { slug } }) => ({
+  titleSuffix: "Event Registration",
+  relativeUrl: `events/${slug}/register`,
+}));
 
 const RegistrationPage = ({
   params: { slug },
