+++ src/app/events/layout.tsx
@@ -1,8 +1,9 @@
 import { genMeta } from "@/utils/route";
 
-export async function generateMetadata() {
-  return genMeta({ titleSuffix: "Events", relativeURL: "/events" });
-}
+export const generateMetadata = genMeta(() => ({
+  titleSuffix: "Events",
+  relativeURL: "/events",
+}));
 
 export default function EventsLayout({ children }) {
   return <>{children}</>;
