+++ src/app/solidity/layout.tsx
@@ -1,8 +1,9 @@
 import { genMeta } from "@/utils/route";
 
-export async function generateMetadata() {
-  return genMeta({ titleSuffix: "Solidity", relativeURL: "/solidity" });
-}
+export const generateMetadata = genMeta(() => ({
+  titleSuffix: "Solidity",
+  relativeURL: "/solidity",
+}));
 
 const SolidityLayout = async ({ children }) => {
   return <>{children}</>;
